cleanup and rename aes ecb and sha256
