+++ lib/util/CMakeLists.txt
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_library(util OBJECT log.cc crypto.cc sha2.cc aesEcb.cc)
+add_library(util OBJECT log.cc crypto.cc sha256.cc aes_ecb.cc)
 target_link_libraries(util crypto zstd)
 
 proofs_add_tests(ceildiv_test)
