+++ lib/util/aes_ecb.cc
@@ -1,27 +1,23 @@
+// This is free and unencumbered software released into the public domain.
+// it comes from https://github.com/kokke/tiny-AES-c/
+// and is stripped down to fit the purpose of longfellow-zk
+
+// This is an implementation of the AES algorithm, specifically ECB
+// mode. Block size is hard-coded to AES128.
+
+// This implementation is verified against the test vectors in:
+//   National Institute of Standards and Technology Special
+//   Publication 800-38A 2001 ED
+
 #if defined(WITHOUT_OPENSSL)
 
-#include <string.h> 
-#include "aes.h"
-
+#include <string.h>
+#include "util/aes_ecb.h"
 
 #define Nb 4
 
-#if defined(AES256) && (AES256 == 1)
-    #define Nk 8
-    #define Nr 14
-#elif defined(AES192) && (AES192 == 1)
-    #define Nk 6
-    #define Nr 12
-#else
-    #define Nk 4        
-    #define Nr 10       
-#endif
-
-
-#ifndef MULTIPLY_AS_A_FUNCTION
-  #define MULTIPLY_AS_A_FUNCTION 0
-#endif
-
+#define Nk 8
+#define Nr 14
 
 typedef uint8_t state_t[4][4];
 
@@ -44,7 +40,6 @@ static const uint8_t sbox[256] = {
   0xe1, 0xf8, 0x98, 0x11, 0x69, 0xd9, 0x8e, 0x94, 0x9b, 0x1e, 0x87, 0xe9, 0xce, 0x55, 0x28, 0xdf,
   0x8c, 0xa1, 0x89, 0x0d, 0xbf, 0xe6, 0x42, 0x68, 0x41, 0x99, 0x2d, 0x0f, 0xb0, 0x54, 0xbb, 0x16 };
 
-#if (defined(CBC) && CBC == 1) || (defined(ECB) && ECB == 1)
 static const uint8_t rsbox[256] = {
   0x52, 0x09, 0x6a, 0xd5, 0x30, 0x36, 0xa5, 0x38, 0xbf, 0x40, 0xa3, 0x9e, 0x81, 0xf3, 0xd7, 0xfb,
   0x7c, 0xe3, 0x39, 0x82, 0x9b, 0x2f, 0xff, 0x87, 0x34, 0x8e, 0x43, 0x44, 0xc4, 0xde, 0xe9, 0xcb,
@@ -62,126 +57,111 @@ static const uint8_t rsbox[256] = {
   0x60, 0x51, 0x7f, 0xa9, 0x19, 0xb5, 0x4a, 0x0d, 0x2d, 0xe5, 0x7a, 0x9f, 0x93, 0xc9, 0x9c, 0xef,
   0xa0, 0xe0, 0x3b, 0x4d, 0xae, 0x2a, 0xf5, 0xb0, 0xc8, 0xeb, 0xbb, 0x3c, 0x83, 0x53, 0x99, 0x61,
   0x17, 0x2b, 0x04, 0x7e, 0xba, 0x77, 0xd6, 0x26, 0xe1, 0x69, 0x14, 0x63, 0x55, 0x21, 0x0c, 0x7d };
-#endif
-
 
 static const uint8_t Rcon[11] = {
   0x8d, 0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80, 0x1b, 0x36 };
 
 #define getSBoxValue(num) (sbox[(num)])
- 
+
 static void KeyExpansion(uint8_t* RoundKey, const uint8_t* Key)
 {
   unsigned i, j, k;
-  uint8_t tempa[4]; 
-  
+  uint8_t tempa[4];
+
   for (i = 0; i < Nk; ++i)
-  {
-    RoundKey[(i * 4) + 0] = Key[(i * 4) + 0];
-    RoundKey[(i * 4) + 1] = Key[(i * 4) + 1];
-    RoundKey[(i * 4) + 2] = Key[(i * 4) + 2];
-    RoundKey[(i * 4) + 3] = Key[(i * 4) + 3];
-  }
+    {
+      RoundKey[(i * 4) + 0] = Key[(i * 4) + 0];
+      RoundKey[(i * 4) + 1] = Key[(i * 4) + 1];
+      RoundKey[(i * 4) + 2] = Key[(i * 4) + 2];
+      RoundKey[(i * 4) + 3] = Key[(i * 4) + 3];
+    }
 
   for (i = Nk; i < Nb * (Nr + 1); ++i)
-  {
-    {
-      k = (i - 1) * 4;
-      tempa[0]=RoundKey[k + 0];
-      tempa[1]=RoundKey[k + 1];
-      tempa[2]=RoundKey[k + 2];
-      tempa[3]=RoundKey[k + 3];
-
-    }
-
-    if (i % Nk == 0)
     {
       {
-        const uint8_t u8tmp = tempa[0];
-        tempa[0] = tempa[1];
-        tempa[1] = tempa[2];
-        tempa[2] = tempa[3];
-        tempa[3] = u8tmp;
+        k = (i - 1) * 4;
+        tempa[0]=RoundKey[k + 0];
+        tempa[1]=RoundKey[k + 1];
+        tempa[2]=RoundKey[k + 2];
+        tempa[3]=RoundKey[k + 3];
+
       }
 
-      {
-        tempa[0] = getSBoxValue(tempa[0]);
-        tempa[1] = getSBoxValue(tempa[1]);
-        tempa[2] = getSBoxValue(tempa[2]);
-        tempa[3] = getSBoxValue(tempa[3]);
-      }
+      if (i % Nk == 0)
+        {
+          {
+            const uint8_t u8tmp = tempa[0];
+            tempa[0] = tempa[1];
+            tempa[1] = tempa[2];
+            tempa[2] = tempa[3];
+            tempa[3] = u8tmp;
+          }
 
-      tempa[0] = tempa[0] ^ Rcon[i/Nk];
+          {
+            tempa[0] = getSBoxValue(tempa[0]);
+            tempa[1] = getSBoxValue(tempa[1]);
+            tempa[2] = getSBoxValue(tempa[2]);
+            tempa[3] = getSBoxValue(tempa[3]);
+          }
+
+          tempa[0] = tempa[0] ^ Rcon[i/Nk];
+        }
+      if (i % Nk == 4)
+        {
+          {
+            tempa[0] = getSBoxValue(tempa[0]);
+            tempa[1] = getSBoxValue(tempa[1]);
+            tempa[2] = getSBoxValue(tempa[2]);
+            tempa[3] = getSBoxValue(tempa[3]);
+          }
+        }
+      j = i * 4; k=(i - Nk) * 4;
+      RoundKey[j + 0] = RoundKey[k + 0] ^ tempa[0];
+      RoundKey[j + 1] = RoundKey[k + 1] ^ tempa[1];
+      RoundKey[j + 2] = RoundKey[k + 2] ^ tempa[2];
+      RoundKey[j + 3] = RoundKey[k + 3] ^ tempa[3];
     }
-#if defined(AES256) && (AES256 == 1)
-    if (i % Nk == 4)
-    {
-      {
-        tempa[0] = getSBoxValue(tempa[0]);
-        tempa[1] = getSBoxValue(tempa[1]);
-        tempa[2] = getSBoxValue(tempa[2]);
-        tempa[3] = getSBoxValue(tempa[3]);
-      }
-    }
-#endif
-    j = i * 4; k=(i - Nk) * 4;
-    RoundKey[j + 0] = RoundKey[k + 0] ^ tempa[0];
-    RoundKey[j + 1] = RoundKey[k + 1] ^ tempa[1];
-    RoundKey[j + 2] = RoundKey[k + 2] ^ tempa[2];
-    RoundKey[j + 3] = RoundKey[k + 3] ^ tempa[3];
-  }
 }
 
 void AES_init_ctx(struct AES_ctx* ctx, const uint8_t* key)
 {
   KeyExpansion(ctx->RoundKey, key);
 }
-#if (defined(CBC) && (CBC == 1)) || (defined(CTR) && (CTR == 1))
-void AES_init_ctx_iv(struct AES_ctx* ctx, const uint8_t* key, const uint8_t* iv)
-{
-  KeyExpansion(ctx->RoundKey, key);
-  memcpy (ctx->Iv, iv, AES_BLOCKLEN);
-}
-void AES_ctx_set_iv(struct AES_ctx* ctx, const uint8_t* iv)
-{
-  memcpy (ctx->Iv, iv, AES_BLOCKLEN);
-}
-#endif
 
 static void AddRoundKey(uint8_t round, state_t* state, const uint8_t* RoundKey)
 {
   uint8_t i,j;
   for (i = 0; i < 4; ++i)
-  {
-    for (j = 0; j < 4; ++j)
     {
-      (*state)[i][j] ^= RoundKey[(round * Nb * 4) + (i * Nb) + j];
+      for (j = 0; j < 4; ++j)
+        {
+          (*state)[i][j] ^= RoundKey[(round * Nb * 4) + (i * Nb) + j];
+        }
     }
-  }
 }
 
 static void SubBytes(state_t* state)
 {
   uint8_t i, j;
   for (i = 0; i < 4; ++i)
-  {
-    for (j = 0; j < 4; ++j)
     {
-      (*state)[j][i] = getSBoxValue((*state)[j][i]);
+      for (j = 0; j < 4; ++j)
+        {
+          (*state)[j][i] = getSBoxValue((*state)[j][i]);
+        }
     }
-  }
 }
 
 static void ShiftRows(state_t* state)
 {
   uint8_t temp;
-  
+
   temp           = (*state)[0][1];
   (*state)[0][1] = (*state)[1][1];
   (*state)[1][1] = (*state)[2][1];
   (*state)[2][1] = (*state)[3][1];
   (*state)[3][1] = temp;
-  
+
   temp           = (*state)[0][2];
   (*state)[0][2] = (*state)[2][2];
   (*state)[2][2] = temp;
@@ -207,34 +187,22 @@ static void MixColumns(state_t* state)
   uint8_t i;
   uint8_t Tmp, Tm, t;
   for (i = 0; i < 4; ++i)
-  {  
-    t   = (*state)[i][0];
-    Tmp = (*state)[i][0] ^ (*state)[i][1] ^ (*state)[i][2] ^ (*state)[i][3] ;
-    Tm  = (*state)[i][0] ^ (*state)[i][1] ; Tm = xtime(Tm);  (*state)[i][0] ^= Tm ^ Tmp ;
-    Tm  = (*state)[i][1] ^ (*state)[i][2] ; Tm = xtime(Tm);  (*state)[i][1] ^= Tm ^ Tmp ;
-    Tm  = (*state)[i][2] ^ (*state)[i][3] ; Tm = xtime(Tm);  (*state)[i][2] ^= Tm ^ Tmp ;
-    Tm  = (*state)[i][3] ^ t ;              Tm = xtime(Tm);  (*state)[i][3] ^= Tm ^ Tmp ;
-  }
+    {
+      t   = (*state)[i][0];
+      Tmp = (*state)[i][0] ^ (*state)[i][1] ^ (*state)[i][2] ^ (*state)[i][3] ;
+      Tm  = (*state)[i][0] ^ (*state)[i][1] ; Tm = xtime(Tm);  (*state)[i][0] ^= Tm ^ Tmp ;
+      Tm  = (*state)[i][1] ^ (*state)[i][2] ; Tm = xtime(Tm);  (*state)[i][1] ^= Tm ^ Tmp ;
+      Tm  = (*state)[i][2] ^ (*state)[i][3] ; Tm = xtime(Tm);  (*state)[i][2] ^= Tm ^ Tmp ;
+      Tm  = (*state)[i][3] ^ t ;              Tm = xtime(Tm);  (*state)[i][3] ^= Tm ^ Tmp ;
+    }
 }
 
-#if MULTIPLY_AS_A_FUNCTION
-static uint8_t Multiply(uint8_t x, uint8_t y)
-{
-  return (((y & 1) * x) ^
-       ((y>>1 & 1) * xtime(x)) ^
-       ((y>>2 & 1) * xtime(xtime(x))) ^
-       ((y>>3 & 1) * xtime(xtime(xtime(x)))) ^
-       ((y>>4 & 1) * xtime(xtime(xtime(xtime(x)))))); /* this last call to xtime() can be omitted */
-  }
-#else
-#define Multiply(x, y)                                \
-      (  ((y & 1) * x) ^                              \
-      ((y>>1 & 1) * xtime(x)) ^                       \
-      ((y>>2 & 1) * xtime(xtime(x))) ^                \
-      ((y>>3 & 1) * xtime(xtime(xtime(x)))) ^         \
-      ((y>>4 & 1) * xtime(xtime(xtime(xtime(x))))))   \
-
-#endif
+#define Multiply(x, y)                              \
+  (  ((y & 1) * x) ^                                \
+     ((y>>1 & 1) * xtime(x)) ^                      \
+     ((y>>2 & 1) * xtime(xtime(x))) ^               \
+     ((y>>3 & 1) * xtime(xtime(xtime(x)))) ^        \
+     ((y>>4 & 1) * xtime(xtime(xtime(xtime(x))))))
 
 
 static void Cipher(state_t* state, const uint8_t* RoundKey)
@@ -244,21 +212,17 @@ static void Cipher(state_t* state, const uint8_t* RoundKey)
   AddRoundKey(0, state, RoundKey);
 
   for (round = 1; ; ++round)
-  {
-    SubBytes(state);
-    ShiftRows(state);
-    if (round == Nr) {
-      break;
+    {
+      SubBytes(state);
+      ShiftRows(state);
+      if (round == Nr) {
+        break;
+      }
+      MixColumns(state);
+      AddRoundKey(round, state, RoundKey);
     }
-    MixColumns(state);
-    AddRoundKey(round, state, RoundKey);
-  }
   AddRoundKey(Nr, state, RoundKey);
 }
-
-
-
-#if defined(ECB) && (ECB == 1)
 
 
 void AES_ECB_encrypt(const struct AES_ctx* ctx, uint8_t* buf)
@@ -267,5 +231,4 @@ void AES_ECB_encrypt(const struct AES_ctx* ctx, uint8_t* buf)
   Cipher((state_t*)buf, ctx->RoundKey);
 }
 
-#endif // #if defined(ECB) && (ECB == 1)
-
+#endif // WITHOUT_OPENSSL
