+++ lib/util/crypto.h
@@ -69,7 +69,7 @@ class SHA256 {
   SHA256_CTX sha_;
 };
 #else
-#include "util/sha2.h" // Replace OpenSSL includes with this
+#include "util/sha256.h" // Replace OpenSSL includes with this
 #include <cstddef> // For size_t
 #include <cstdint> // For uint8_t, uint64_t
 #include <cstring> // For memcpy
@@ -190,10 +190,8 @@ class PRF {
   EVP_CIPHER_CTX* ctx_;
 };
 #else
-#include "util/aesEcb.h" // Replace OpenSSL includes with this
-#include <cstring> 
-#undef AES128
-#define AES256 1
+#include "util/aes_ecb.h" // Replace OpenSSL includes with this
+#include <cstring>
 
 class PRF {
   public:
@@ -201,35 +199,35 @@ class PRF {
    static constexpr size_t kPRFKeySize = 32;    // AES-256 key size (32 bytes)
    static constexpr size_t kPRFInputSize = 16;   // AES block size for input (16 bytes)
    static constexpr size_t kPRFOutputSize = 16;  // AES block size for output (16 bytes)
- 
+
    // Constructor - takes 32-byte key for AES-256
    explicit PRF(const uint8_t key[kPRFKeySize]) {
      AES_init_ctx(&ctx_, key);  // Initialize with key
    }
- 
+
    // Destructor - no cleanup needed for stack-allocated ctx_
    ~PRF() = default;
- 
+
    // Delete copy constructor and assignment operator for safety
    PRF(const PRF&) = delete;
    PRF& operator=(const PRF&) = delete;
- 
+
    // Evaluates the PRF (pseudorandom function) on the input
    // This performs AES-256 encryption in ECB mode on the input block
-   
+
    void Eval(uint8_t out[kPRFOutputSize], const uint8_t in[kPRFInputSize]) {
      // Copy input to output (AES_ECB_encrypt works in-place)
     memcpy(out, in, kPRFInputSize);
-    
+
     // Perform AES-256 ECB encryption
     AES_ECB_encrypt(&ctx_, out);
-    
+
     // Result is now in out buffer
   }
   private:
    AES_ctx ctx_;  // AES context that holds the expanded key schedule
  };
- 
+
 #endif
 
 // Generate n random bytes, following the openssl API convention.
